+++ NEWS.md
@@ -8,6 +8,7 @@
 * Added parameter to disable punctuation analysis.
 * Added convenience parameter simple_count to count word presences.
 * Add support for multi-word modifiers.
+* Add support for multi-word negations.
 * Added unit tests for all new features.
 * Internal code cleanup.